requirements.txt
cryptography==45.0.5
docx2txt==0.9
Flask==3.1.1
PyPDF2==3.0.1
python-docx==1.2.0  
scikit-learn==1.7.1
